DRY Refactored Oracle Java and OpenJDK helper to put common functionality into common JDK helper
